automate articles

+++ .buildscript/jekyll.thor
@@ -6,15 +6,7 @@ require "net/http"
 class Jekyll < Thor
   desc "new", "create a new post"
   method_option :editor, :default => "subl"
-  def new(*title)
-    title = title.join(" ")
-    date = Time.now.strftime("%Y-%m-%d")
-    filename = "_posts/#{date}-#{title.to_url}.markdown"
-
-    if File.exist?(filename)
-      abort("#{filename} already exists!")
-    end
-
+  def new()
     url = "https://newsapi.org/v1/articles?source=breitbart-news&sortBy=top&apiKey=63c394c42ffd45afab11bb4928a4b879"
     uri = URI(url)
     response = Net::HTTP.get(uri)
@@ -24,6 +16,14 @@ class Jekyll < Thor
     description = json["description"]
     url = json["url"]
     urlToImage = json["urlToImage"]
+
+    title = title.gsub(" ", "-")
+    date = Time.now.strftime("%Y-%m-%d")
+    filename = "../_posts/#{date}-#{title.to_url}.markdown"
+
+    if File.exist?(filename)
+      abort("#{filename} already exists!")
+    end
 
     puts "Creating new post: #{filename}"
     open(filename, "w") do |post|
